% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_problem.R
\name{abcost}
\alias{abcost}
\title{Compute cost matrix for label pairs}
\usage{
abcost(a, b, .unpaired = 2 * length(a), .names = TRUE)
}
\arguments{
\item{a}{Reference labels}

\item{b}{(Partially) Paired labels. Inputs \code{a} and \code{b}
must be of equal length}

\item{.unpaired}{Cost associated with assigning a
non-overlapping pair of labels}

\item{.names}{Include table row/column names}
}
\value{
Square cost matrix where the columns are associated with
  the reference labels, \code{a}, and the rows are associated with
  the compared labels, \code{b}
}
\description{
Input \code{a} taken as the reference lables.
Implementation uses \code{base::table}.
}
\examples{
x <- sample.int(10, 100, replace = TRUE)  # Reference labels
map <- sample(1:10)
y <- map[x]                               # Renamed labels
abcost(x, y)

}
\seealso{
\code{\link[base]{table}}
}
