% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_problem.R
\name{assignment.cost}
\alias{assignment.cost}
\title{Compute the cost of a given labeling}
\usage{
assignment.cost(cost, label)
}
\arguments{
\item{cost}{Square cost matrix}

\item{label}{Integer labeling. Should be some permutation of
\code{[1, ..., nrow(cost)]}.}
}
\value{
Cost associated with the labeling (scalar)
}
\description{
Compute the cost of a given labeling
}
\examples{
C <- simulate.difficult.cost(10)  # Cost matrix
f <- ctc(C)
assignment.cost(C, f) == attr(f, "cost")  # [1] TRUE

}
\seealso{
\code{\link{ctc}}, \code{\link{simulate.difficult.cost}}
}
