% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment1-package.R
\docType{package}
\name{assignment1-package}
\alias{assignment1}
\alias{assignment1-package}
\title{assignment1: Data Package With Data from Digital Word Recognition by Howes and Solomon}
\description{
The package provides a tidy and clean dataset that can be used for Assignment 1 of the course CGSC 5901/4900 X Special Topics in Cognitive Science: Advanced Statistics for Cognitive Science.
}
\author{
\strong{Maintainer}: Romeo Penheiro \email{romeoalexp@gmail.com} (\href{https://orcid.org/0000-0001-5267-9121}{ORCID})

}
\keyword{internal}
