% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlogram.R
\name{correlogram}
\alias{correlogram}
\title{Correlogram Function
This function display the correlogram of all numeric variables of input year.}
\usage{
correlogram(data, year)
}
\arguments{
\item{year}{Which is your year of interest? Range from 1950 to 2017.}
}
\description{
Correlogram Function
This function display the correlogram of all numeric variables of input year.
}
\examples{
correlogram()
}
\keyword{correlogram}
