% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_problem.R
\name{ctc}
\alias{ctc}
\title{Solve the Assignment Problem}
\usage{
ctc(cost, .tol = NA)
}
\arguments{
\item{cost}{Square cost matrix}

\item{.tol}{Tolerance for equivalence of non-integer costs.
If not provided, costs are coerced to integer values}
}
\value{
The map \eqn{f(A)}
}
\description{
Given a cost function \eqn{C(a, b)}, \eqn{a \in A} and
\eqn{b \in B}, the "Assignment Problem" is to find a bijective map
\eqn{f : A \to B} that minimizes,
\deqn{\sum_{a \in A} C(a, f(a)).}{\sum_a C(a, f(a)).}
In the solution, we assume that A is the index set of the *columns*
of the cost matrix \eqn{C}.
}
\examples{
C <- simulate.difficult.cost(10)  # Cost matrix
ctc(C)

}
\references{
Carpaneto, Giorgio, and Paolo Toth.
"Primal-dual algrorithms for the assignment problem."
Discrete Applied Mathematics 18.2 (1987): 137-153.
(\href{https://www.sciencedirect.com/science/article/pii/0166218X87900163}{sciencedirect.com})
}
\seealso{
\code{\link{abcost}}, \code{\link{simulate.difficult.cost}}
}
