% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Function: draw_a_name
This function returns the graph of the evolution of the occurence of a name over time in France}
\usage{
draw_a_name(the_sex, the_name)
}
\arguments{
\item{the_sex}{sex of the person}

\item{the_name}{name of the person}
}
\value{
graph of the evolution of name over time
}
\description{
Function: draw_a_name
This function returns the graph of the evolution of the occurence of a name over time in France
}
\examples{
draw_a_name(the_sex = "M", the_name = "Pierre")

}
