% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names_dygraph}
\alias{draw_names_dygraph}
\title{Function: draw_names_dygraph
This function plots the occurence of two names over time}
\usage{
draw_names_dygraph(names)
}
\arguments{
\item{names}{noms recherchés}
}
\value{
graph
}
\description{
Function: draw_names_dygraph
This function plots the occurence of two names over time
}
\examples{
draw_names_dygraph(names = "Pierre")

}
