% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eighty.R
\name{eighty}
\alias{eighty}
\title{Calculate subsidy amount to be returned on 80\% subsidy}
\usage{
eighty(x)
}
\arguments{
\item{x}{numeric vector. Total cost or baseline market value before application of subsidy.}
}
\value{
A vector containing the subsidy amount to be returned on total cost based on 80\% subsidy
}
\description{
Calculate subsidy amount to be returned on 80\% subsidy
}
\examples{
eighty(1500)
}
