% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fifty.R
\name{fifty}
\alias{fifty}
\title{Calculate subsidy amount to be returned on 50\% subsidy}
\usage{
fifty(x)
}
\arguments{
\item{x}{numeric vector. Total cost or baseline market value before application of subsidy.}
}
\value{
A vector containing the subsidy amount to be returned on total cost based on 50\% subsidy
}
\description{
Calculate subsidy amount to be returned on 50\% subsidy
}
\examples{
fifty(1200)
}
