% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_function.R
\name{linear_model}
\alias{linear_model}
\title{a linear model function mimicking lm}
\usage{
linear_model(form, data, contrasts = NULL)
}
\arguments{
\item{form}{A formula}

\item{raw_data}{A data frame}

\item{const}{A list of contrasts for factor variables}
}
\description{
a linear model function mimicking lm
}
\examples{



}
