% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_function.R
\name{linear_model2}
\alias{linear_model2}
\title{A linear model function which solves using gradient descent}
\usage{
linear_model2(form, data, contrasts = NULL)
}
\arguments{
\item{form}{A formula}

\item{data}{A data frame}

\item{const}{(optional) A list of contrasts for factor variables}
}
\value{
A list of coefficients for \code{data} in the formula \code{form}
}
\description{
A linear model function which solves using gradient descent
}
\examples{
linear_model2(Sepal.Length ~ ., iris)
linear_model2(Sepal.Length ~ ., iris, contrasts = list(Species = "contr.sum"))

}
