% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_patho}
\alias{lm_patho}
\title{an ex}
\format{A data frame with 2 rows and 3 variables:
\describe{
  \item{x1}
  \item{x2}
  ...
}}
\source{
BIS557/homework-1-cck42
}
\usage{
lm_patho
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
