% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_pts.R
\name{max_pts}
\alias{max_pts}
\title{Maximum Points Function
This function allows you to get the player earned the most points for a specific year.}
\usage{
max_pts(data, year)
}
\arguments{
\item{year}{Which is your year of interest? Range from 1950 to 2017.}
}
\description{
Maximum Points Function
This function allows you to get the player earned the most points for a specific year.
}
\examples{
max_pts()
}
\keyword{points}
