% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{My CSV reader
This function reads CSV files in a folder}
\usage{
my_csv_reader(folder)
}
\arguments{
\item{folder}{containing the .csv files}
}
\value{
the dataframe contained in the csv
}
\description{
My CSV reader
This function reads CSV files in a folder
}
\examples{
\dontrun{
my_csv_reader("Desktop/jeux_video.csv")
}
}
