% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player Function
This function get the oldest player of the input year, if multiple players found, return the one earned the most points}
\usage{
oldest_player(data, year)
}
\arguments{
\item{year}{Which is your year of interest? Range from 1950 to 2017.}
}
\description{
Oldest Player Function
This function get the oldest player of the input year, if multiple players found, return the one earned the most points
}
\examples{
oldest_player()
}
\keyword{age}
