% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{read_excel_multi}
\alias{read_excel_multi}
\title{Function: multi_excel
This function imports all sheets from an xlsx file}
\usage{
read_excel_multi(file)
}
\arguments{
\item{file}{the xlsx file from which you want to extract sheets}
}
\value{
imports sheets
}
\description{
Function: multi_excel
This function imports all sheets from an xlsx file
}
\examples{
\dontrun{
read_excel_multi("xlsxfile")
}


}
