% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_problem.R
\name{relabel}
\alias{relabel}
\title{Relabel \code{x} to match \code{reference} based on minimum cost
solution}
\usage{
relabel(x, reference)
}
\arguments{
\item{x}{Set of labels to be recoded}

\item{reference}{Set of reference labels to match}
}
\value{
\code{x}, but relabeld to match the values in
        \code{reference} as closely as possible
}
\description{
Relabel \code{x} to match \code{reference} based on minimum cost
solution
}
