% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save as CSV
This function saves the dataframe as a csv}
\usage{
save_as_csv(dframe, filename, row.names = FALSE, ...)
}
\arguments{
\item{dframe}{the dataframe you want to save as .csv}

\item{filename}{the name and path you want to for your new .csv file}

\item{row.names}{false}
}
\value{
the file
}
\description{
Save as CSV
This function saves the dataframe as a csv
}
\examples{
save_as_csv (titanic, titanic.csv, row.names = FALSE, ...)
}
