% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_problem.R
\name{simulate.difficult.cost}
\alias{simulate.difficult.cost}
\title{Simulate difficult to solve cost matrix}
\usage{
simulate.difficult.cost(n)
}
\arguments{
\item{n}{Dimension of simulated cost matrix}
}
\value{
\code{n x n} cost matrix
}
\description{
Simulate difficult to solve cost matrix
}
