% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_problem.R
\name{ctc}
\alias{ctc}
\title{Solve the Assignment Problem}
\usage{
ctc(cost, .eps = 1e-06)
}
\arguments{
\item{cost}{Square cost matrix.}

\item{.eps}{Tolerance for equivalence of non-integer costs.
That is, an \eqn{\epsilon > 0} such that two values
\eqn{x} and \eqn{y} can be taken as equivalent for
practical purposes when \eqn{|x - y| < \epsilon}.
Unused/not needed if \code{cost} is a matrix of integers.}
}
\value{
The map \eqn{f(A)}
}
\description{
Given a cost function \eqn{C(a, b)}, \eqn{a \in A} and
\eqn{b \in B}, the "Assignment Problem" is to find a bijective map
\eqn{f : A \to B} that minimizes,
\deqn{\sum_{a \in A} C(a, f(a)).}{\sum_a C(a, f(a)).}
In the solution, we assume that A is the index set of the *columns*
of the cost matrix \eqn{C}.
}
\examples{
C <- simulate.difficult.cost(10)  # Cost matrix
map <- ctc(C)
sum(diag( C[map, ] )) == attr(map, "cost")  # TRUE

}
\references{
Carpaneto, Giorgio, and Paolo Toth.
"Primal-dual algrorithms for the assignment problem."
Discrete Applied Mathematics 18.2 (1987): 137-153.
(\href{https://www.sciencedirect.com/science/article/pii/0166218X87900163}{sciencedirect.com})
}
\seealso{
\code{\link{abcost}}, \code{\link{simulate.difficult.cost}}
}
