% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{Function: my_cv_reader
This function takes all the csv from a folder, imports them, puts them in a list and returns the list}
\usage{
my_csv_reader(folder)
}
\arguments{
\item{folder}{the folder where the csv are}
}
\value{
list of the csv
}
\description{
Function: my_cv_reader
This function takes all the csv from a folder, imports them, puts them in a list and returns the list
}
\examples{

\dontrun{
my_csv_reader("cheminversundossier")
}


}
